dev: fix some bugs in comments routers
